Moved "Import Associates" modal to accounts page, fixed side menu issue
