+++ src/containers/b2b/B2BMain.less
@@ -198,6 +198,7 @@
         position: static;
         width: 200px;
         margin-right: 40px;
+        margin-top: 40px;
       }
     }
